calculator moved
